+++ R/plotBoxEllipses.R
@@ -36,6 +36,8 @@ bdpPlotBE<-function(bdpreg,cov.method="MCD",ci=0.95,...){
     het.text <- "Linear heteroscedastic model"
   } else if (dat$heteroscedastic == "exponential") {
     het.text <- "Exponential heteroscedastic model"
+  } else if (dat$heteroscedastic == "binomial"){
+    het.text <- "Binomial error structure model"
   } else {
     het.text <- "Homoscedastic model"
   }