+++ src/RcppExports.cpp
@@ -12,6 +12,7 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 
+RcppExport SEXP _rcpp_module_boot_stan_fit4bdpreg_binomtrunc_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4bdpreg_exphettrunc_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4bdpreg_homo_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4bdpreg_homotrunc_mod();
@@ -19,6 +20,7 @@ RcppExport SEXP _rcpp_module_boot_stan_fit4bdpreg_linhet_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4bdpreg_linhettrunc_mod();
 
 static const R_CallMethodDef CallEntries[] = {
+    {"_rcpp_module_boot_stan_fit4bdpreg_binomtrunc_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4bdpreg_binomtrunc_mod, 0},
     {"_rcpp_module_boot_stan_fit4bdpreg_exphettrunc_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4bdpreg_exphettrunc_mod, 0},
     {"_rcpp_module_boot_stan_fit4bdpreg_homo_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4bdpreg_homo_mod, 0},
     {"_rcpp_module_boot_stan_fit4bdpreg_homotrunc_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4bdpreg_homotrunc_mod, 0},